fix : 현재 위치 검색 진행중

+++ src/main/java/com/ssafy/trip/spot/dto/response/NearbySearchResponse.java
@@ -3,12 +3,13 @@ package com.ssafy.trip.spot.dto.response;
 import com.ssafy.trip.spot.dto.SpotDto;
 
 import java.util.List;
+import software.amazon.awssdk.services.s3.endpoints.internal.Value.Int;
 
 public class NearbySearchResponse {
     private List<SpotDto> spots;
-    private Long nextCursor;
+    private Integer nextCursor;
 
-    public NearbySearchResponse(List<SpotDto> spots, Long nextCursor){
+    public NearbySearchResponse(List<SpotDto> spots, Integer nextCursor){
         this.spots = spots;
         this.nextCursor = nextCursor;
     }
